uav_project/requirements.txt
# Core Dependencies
dronekit>=2.9.2
pymavlink>=2.4.37
pyserial>=3.5

# GUI Framework
PyQt5>=5.15.0

# Computer Vision
opencv-contrib-python>=4.8.0
torch>=1.13.0
ultralytics>=8.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Networking
websockets>=10.0
aiohttp>=3.8.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Logging and Monitoring
structlog>=22.0.0

# Utilities
psutil>=5.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
coverage>=6.0.0
